[Feat]: Replace loopback with 0.0.0.0
This commit replaces the 127.0.0.1 address with 0.0.0.0, for ease of use. It is unlikely that the end user running the API on the Pi is also using the Pi to run the frontend (and it makes things easier for me to develop this way. I should consider containerization or python-dotenv, though.)