[Chore] Update requirements
I manually added three libraries here, since the gpio library is not always installed by default, the adafruit one definitely is not, and the standard tag for uvicorn is needed to have websocket support. I have not tested these yet, and did not generate the requirements file on my testing Pi. Oops.